# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 심어봄 API
  version: 1.0.0
  description: |
    현재 Laravel 12 백엔드에 구현되어 테스트되는 심어봄 API 계약입니다.
    계획 단계의 엔드포인트는 포함하지 않습니다.
servers:
  - url: http://localhost:8000/api/v1
    description: Laravel 로컬 개발 서버
...
paths:
...
  /watering-schedules/{wateringScheduleId}/complete:
    post:
      tags: [Watering]
      summary: 물주기 완료 및 다음 일정 계산
      operationId: completeWatering
      parameters:
        - $ref: '#/components/parameters/WateringScheduleId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteWateringInput' }
      responses:
        '201':
          description: 완료 기록 생성과 다음 일정 계산 성공
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompleteWateringResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '419': { $ref: '#/components/responses/CsrfMismatch' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /watering-schedules/{wateringScheduleId}/snoozes:
    get:
      tags: [Watering]
      summary: 물주기 미루기 이력 조회
      operationId: listWateringSnoozes
      parameters:
        - $ref: '#/components/parameters/WateringScheduleId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: 미루기 이력 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/WateringSnooze' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Watering]
      summary: 다음 물주기 시각 미루기
      operationId: snoozeWatering
      parameters:
        - $ref: '#/components/parameters/WateringScheduleId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnoozeWateringInput' }
      responses:
        '201':
          description: 물주기 미루기 성공
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnoozeWateringResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '419': { $ref: '#/components/responses/CsrfMismatch' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Laravel Sanctum SPA 세션 쿠키. 실제 쿠키 이름은 배포 설정으로 변경할 수 있습니다.
  headers:
    ETag:
      description: 현재 자원의 정수 version을 따옴표로 감싼 값
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
        examples: ['"3"']
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    WateringScheduleId:
      name: wateringScheduleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 마지막으로 조회한 자원의 ETag 값
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
        examples: ['"3"']
...
  responses:
    Unauthenticated:
      description: 인증이 필요하거나 토큰이 만료됨
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 자원에 접근할 권한이 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 자원을 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 현재 자원 상태 또는 관계와 요청이 충돌함
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: 입력값 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionFailed:
      description: If-Match 버전이 현재 자원의 버전과 달라 변경을 적용하지 않음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CsrfMismatch:
      description: CSRF 쿠키 또는 요청 헤더 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    WateringSchedule:
      type: object
      required:
        - id
        - seasonId
        - cropId
        - intervalDays
        - nextWateringAt
        - enabled
        - version
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        seasonId: { type: string, format: uuid }
        cropId: { type: string }
        intervalDays: { type: integer, minimum: 1, maximum: 365 }
        nextWateringAt: { type: string, format: date-time }
        enabled: { type: boolean }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CompleteWateringInput:
      type: object
      additionalProperties: false
      required: [wateredAt, amountMl, memo]
      properties:
        wateredAt:
          type: string
          format: date-time
          description: 재배 시즌 안의 실제 완료 시각
        amountMl:
          type: [integer, 'null']
          minimum: 1
          maximum: 100000
        memo:
          type: [string, 'null']
          maxLength: 500
    WateringLog:
      type: object
      required: [id, scheduleId, userId, scheduledFor, wateredAt, amountMl, memo, createdAt]
      properties:
        id: { type: string, format: uuid }
        scheduleId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        scheduledFor: { type: string, format: date-time }
        wateredAt: { type: string, format: date-time }
        amountMl: { type: [integer, 'null'] }
        memo: { type: string }
        createdAt: { type: string, format: date-time }
    CompleteWateringResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [schedule, log]
          properties:
            schedule: { $ref: '#/components/schemas/WateringSchedule' }
            log: { $ref: '#/components/schemas/WateringLog' }
    SnoozeWateringInput:
      type: object
      additionalProperties: false
      required: [snoozedUntil]
      properties:
        snoozedUntil:
          type: string
          format: date-time
          description: 현재 예정 시각보다 늦고 재배 시즌 안인 시각
    WateringSnooze:
      type: object
      required: [id, scheduleId, originalAt, snoozedUntil, createdAt]
      properties:
        id: { type: string, format: uuid }
        scheduleId: { type: string, format: uuid }
        originalAt: { type: string, format: date-time }
        snoozedUntil: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    SnoozeWateringResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [schedule, snooze]
          properties:
            schedule: { $ref: '#/components/schemas/WateringSchedule' }
            snooze: { $ref: '#/components/schemas/WateringSnooze' }
    PaginationMeta:
      type: object
      required: [currentPage, perPage, total, lastPage]
      properties:
        currentPage:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        lastPage:
          type: integer
          minimum: 1
    PaginatedResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [VALIDATION_FAILED]
            message:
              type: string
              examples: [입력값을 확인해 주세요.]
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
security:
  - sessionCookie: []